test: fix failing logging tests
